backend/requirements.txt
flask==3.0.3
langchain==0.2.11
google-generativeai==0.7.2
pymongo[srv]==3.12
python-dotenv==1.0.1
flask-cors==4.0.1
